separated video upload feature into ExerciseBox made it into button on FeedbackPanel
